backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
aiohttp
numpy

scikit-learn
joblib
pandas
rasterio
shapely